+++ sphinxcontrib/osint/scripts/text.py
@@ -8,13 +8,13 @@ The text scripts
 from __future__ import annotations
 import os
 import sys
+import time
 from datetime import date
 import json
 import click
 
 from ..plugins.text import Text
-from . import parser_makefile, cli, get_app
-
+from . import parser_makefile, cli, get_app, load_quest
 
 __author__ = 'bibi21000 aka Sébastien GALLET'
 __email__ = 'bibi21000@gmail.com'
@@ -22,9 +22,10 @@ __email__ = 'bibi21000@gmail.com'
 
 @cli.command()
 @click.option('--delete/--no-delete', default=True, help="Delete file in text_cache")
+@click.option('--html/--no-html', default=False, help="File contains html data")
 @click.argument('textfile', default=None)
 @click.pass_obj
-def store(common, delete, textfile):
+def store(common, delete, html, textfile):
     """Import text in store"""
     sourcedir, builddir = parser_makefile(common.docdir)
     app = get_app(sourcedir=sourcedir, builddir=builddir)
@@ -36,26 +37,42 @@ def store(common, delete, textfile):
     with open(textfile, 'r') as f:
         text = f.read()
 
-    result = {
-      "title": None,
-      "author": 'osint_import_text',
-      "hostname": None,
-      "date": None,
-      "fingerprint": None,
-      "id": None,
-      "license": None,
-      "comments": "",
-      "text": text,
-      "language": None,
-      "image": None,
-      "pagetype": None,
-      "filedate": date.today().isoformat(),
-      "source": None,
-      "source-hostname": None,
-      "excerpt": None,
-      "categories": None,
-      "tags": None,
-    }
+    head = str(text[:30]).lower().replace(' ','').replace('\n','')
+    if html is True and head.startswith("<!doctypehtml>") is False:
+        click.echo('Seem not an HTML file : %s' % text[:30])
+        sys.exit(2)
+
+    if html is False and head.startswith("<!doctypehtml>") is True:
+        click.echo('Seem an HTML file : %s' % text[:30])
+        sys.exit(2)
+
+    if html is False:
+
+        result = {
+          "title": None,
+          "author": 'osint_import_text',
+          "hostname": None,
+          "date": None,
+          "fingerprint": None,
+          "id": None,
+          "license": None,
+          "comments": "",
+          "text": text,
+          "language": None,
+          "image": None,
+          "pagetype": None,
+          "filedate": date.today().isoformat(),
+          "source": None,
+          "source-hostname": None,
+          "excerpt": None,
+          "categories": None,
+          "tags": None,
+        }
+
+    else:
+
+        from trafilatura import extract
+        result = Text.traf_extract(text)
 
     Text.update(app, result, textfile)
 
@@ -67,3 +84,30 @@ def store(common, delete, textfile):
         cachef = os.path.join(sourcedir, app.config.osint_text_cache, os.path.splitext(os.path.basename(textfile))[0] + '.json')
         if os.path.isfile(cachef):
             os.remove(cachef)
+
+@cli.command()
+@click.argument('url', default=None)
+@click.option('--before', default=28800, help="Number of seconds the file has not been modified")
+@click.pass_obj
+def refresh(common, url, before):
+    """Refresh text from site url (ie wikipedia.org)"""
+    sourcedir, builddir = parser_makefile(common.docdir)
+    app = get_app(sourcedir=sourcedir, builddir=builddir)
+
+    if app.config.osint_text_enabled is False:
+        print('Plugin text is not enabled')
+        sys.exit(1)
+
+    if url is None or len(url) < 7:
+        print('URL too short : %r' % url)
+        sys.exit(2)
+
+    data = load_quest(builddir)
+
+    Text.init(app)
+
+    for src in data.sources:
+        if data.sources[src].url is not None and url in data.sources[src].url:
+            print(data.sources[src].name, data.sources[src].url)
+            Text.save(app, data.sources[src].name, data.sources[src].url, update=True, before=time.time() - before)
+            time.sleep(2)